feat: add shared file deletion helper
